requirements.txt
Django==2.1.3
psycopg2
djangorestframework==3.9.0
